BIG: Implement append, exists_and_return, new format; REMOVE OLD DB as well :tada:

+++ src/algo_loc.rs
@@ -1,5 +1,5 @@
 // use crate::db::DB;
-use crate::db_new::DB;
+use crate::db::DB;
 use crate::git_command_algo::extract_details;
 use std::collections::HashMap;
 
@@ -64,8 +64,8 @@ pub fn get_unique_files_changed(
                 let mut res_string: String = res;
                 for each_unvisited_index in unvisited_indices {
                     let output_single_line = perform_for_single_line(
-                        each_unvisited_index,
-                        each_unvisited_index,
+                        each_unvisited_index as usize,
+                        each_unvisited_index as usize,
                         origin_file_path.clone(),
                         db_obj,
                         false,
@@ -88,8 +88,8 @@ pub fn get_unique_files_changed(
             let mut final_result = "".to_string();
             for each_unvisited_index in unvisited_indices {
                 let output_single_line = perform_for_single_line(
-                    each_unvisited_index,
-                    each_unvisited_index,
+                    each_unvisited_index as usize,
+                    each_unvisited_index as usize,
                     origin_file_path.clone(),
                     db_obj,
                     false,
@@ -196,8 +196,8 @@ pub fn get_contextual_authors(
                 let mut res_string: String = res;
                 for each_unvisited_index in unvisited_indices {
                     let output_single_line = perform_for_single_line(
-                        each_unvisited_index,
-                        each_unvisited_index,
+                        each_unvisited_index as usize,
+                        each_unvisited_index as usize,
                         origin_file_path.clone(),
                         db_obj,
                         true,
@@ -220,8 +220,8 @@ pub fn get_contextual_authors(
             let mut final_result = "".to_string();
             for each_unvisited_index in unvisited_indices {
                 let output_single_line = perform_for_single_line(
-                    each_unvisited_index,
-                    each_unvisited_index,
+                    each_unvisited_index as usize,
+                    each_unvisited_index as usize,
                     origin_file_path.clone(),
                     db_obj,
                     true,
